requirements.txt
beautifulsoup4==4.3.2
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-crispy-forms==1.4.0
django-durationfield==0.5.1
django-heroku==0.3.1
django-jet==1.0.8
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
lxml==4.4.1
psycopg2==2.8.3
python-dateutil==2.4.2
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tldextract==1.6
unicodecsv==0.13.0
urllib3==1.25.3
whitenoise==4.1.3
